New urxvt font + color scheme.
Basically a mix of URxvt configs I found on the archlinux forums. Going
to try this setup for a bit and see if I can read it for a few hours at
a time without getting a headache.

I also changed the *ugly* mode line colors in .vim/colors/ir_black.vim.